+++ examples/basic/src/app/robots.ts
@@ -1,0 +1,7 @@
+import { createNextRobots } from "opensitestack/next";
+
+import { getCurrentSite } from "@/lib/current-site";
+
+export default async function robots() {
+  return createNextRobots(await getCurrentSite());
+}